Add initial page for initial game controls
Main.cpp is changed to Game.cpp. Now initial page of the game with Play button as added to the main. More buttons and sound will be added to main including saved levels and etc. A customised button is added with a custom texture. status of hover and pressed are also added.